Utility folders created pt. 2
